functional lab 4

+++ AIMR_source_code/run/lab4.h
@@ -3,8 +3,8 @@
 
 void GoToRules(float xt, float yt);
 void GoTo_FRB(float xt, float yt);
-int ResponseToVel(float response);
-int ResponseToRot(float response);
+double ResponseToVel(float response);
+double ResponseToRot(float response);
 void printSets();
 
 void lab4();
